+++ parsley-core/src/ghc-8.10+/Parsley/Internal/Backend/Machine/BindingOps.hs
@@ -27,9 +27,9 @@ import Data.Array.Unboxed                              (UArray)
 import Data.ByteString.Internal                        (ByteString)
 import Data.Text                                       (Text)
 import Parsley.Internal.Backend.Machine.InputRep       (Rep)
-import Parsley.Internal.Backend.Machine.Types.Base     (Handler#)
+import Parsley.Internal.Backend.Machine.Types.Base     (Handler#, Line, Col)
 import Parsley.Internal.Backend.Machine.Types.Dynamics (DynSubroutine, DynCont, DynHandler)
-import Parsley.Internal.Backend.Machine.Types.Statics  (StaCont#, StaHandler#)
+import Parsley.Internal.Backend.Machine.Types.Statics  (StaCont#, StaHandler#, StaSubroutine#, Input#(..))
 import Parsley.Internal.Common.Utils                   (Code)
 import Parsley.Internal.Core.InputTypes                (Text16, CharList, Stream)
 
@@ -60,13 +60,14 @@ class HandlerOps o where
                -> (DynHandler s o a -> Code b) -- ^ The continuation that expects the bound handler
                -> Code b
 
-#define deriveHandlerOps(_o)                    \
-instance HandlerOps _o where                    \
-{                                               \
-  bindHandler# h k = [||                        \
-    let handler (o# :: Rep _o) = $$(h [||o#||]) \
-    in $$(k [||handler||])                      \
-  ||];                                          \
+#define deriveHandlerOps(_o)                                 \
+instance HandlerOps _o where                                 \
+{                                                            \
+  bindHandler# h k = [||                                     \
+    let handler (line :: Line) (col :: Col) (o# :: Rep _o) = \
+          $$(h (Input# [||o#||] [||line||] [||col||]))       \
+    in $$(k [||handler||])                                   \
+  ||];                                                       \
 };
 inputInstances(deriveHandlerOps)
 
@@ -85,11 +86,12 @@ class JoinBuilder o where
                   -> (DynCont s o a x -> Code b) -- ^ The continuation that expects the bound join point
                   -> Code b
 
-#define deriveJoinBuilder(_o)                                                             \
-instance JoinBuilder _o where                                                             \
-{                                                                                         \
-  setupJoinPoint# binding k =                                                             \
-    [|| let join x !(o# :: Rep _o) = $$(binding [||x||] [||o#||]) in $$(k [||join||]) ||] \
+#define deriveJoinBuilder(_o)                                                                    \
+instance JoinBuilder _o where                                                                    \
+{                                                                                                \
+  setupJoinPoint# binding k =                                                                    \
+    [|| let join x (line :: Line) (col :: Col) !(o# :: Rep _o) =                                 \
+              $$(binding [||x||] (Input# [||o#||] [||line||] [||col||])) in $$(k [||join||]) ||] \
 };
 inputInstances(deriveJoinBuilder)
 
@@ -106,8 +108,8 @@ class RecBuilder o where
 
   @since 1.4.0.0
   -}
-  bindIterHandler# :: (Code (Rep o) -> StaHandler# s o a)        -- ^ The iter handler to bind
-                   -> (Code (Rep o -> Handler# s o a) -> Code b) -- ^ The continuation that accepts the bound handler
+  bindIterHandler# :: (Input# o -> StaHandler# s o a)                           -- ^ The iter handler to bind
+                   -> (Code (Line -> Col -> Rep o -> Handler# s o a) -> Code b) -- ^ The continuation that accepts the bound handler
                    -> Code b
 
   {-|
@@ -115,31 +117,33 @@ class RecBuilder o where
 
   @since 1.4.0.0
   -}
-  bindIter# :: Code (Rep o)                                                -- ^ Initial offset for the loop.
-            -> (DynHandler s o a -> Code (Rep o) -> Code (ST s (Maybe a))) -- ^ The code for the loop given handler and offset.
-            -> Code (ST s (Maybe a))                                       -- ^ Code of the executing loop.
+  bindIter# :: Input# o                                                                             -- ^ Initial offset for the loop.
+            -> (Code (Line -> Col -> Rep o -> ST s (Maybe a)) -> Input# o -> Code (ST s (Maybe a))) -- ^ The code for the loop given self-call and offset.
+            -> Code (ST s (Maybe a))                                                                -- ^ Code of the executing loop.
 
   {-|
   Creates a binding for a regular let-bound parser.
 
   @since 1.4.0.0
   -}
-  bindRec#  :: (DynSubroutine s o a x -> DynCont s o a x -> Code (Rep o) -> DynHandler s o a -> Code (ST s (Maybe a))) -- ^ Code for the binding, accepting itself as an argument.
-            -> DynSubroutine s o a x                                                                                   -- ^ The code that represents this binding's name.
+  bindRec#  :: (DynSubroutine s o a x -> StaSubroutine# s o a x) -- ^ Code for the binding, accepting itself as an argument.
+            -> DynSubroutine s o a x                             -- ^ The code that represents this binding's name.
 
-#define deriveRecBuilder(_o)                                                                           \
-instance RecBuilder _o where                                                                           \
-{                                                                                                      \
-  bindIterHandler# h k = [||                                                                           \
-      let handler (c# :: Rep _o) (o# :: Rep _o) = $$(h [||c#||] [||o#||]) in $$(k [||handler||])       \
-    ||];                                                                                               \
-  bindIter# o l = [||                                                                                  \
-      let loop !(o# :: Rep _o) = $$(l [||loop||] [||o#||])                                             \
-      in loop $$o                                                                                      \
-    ||];                                                                                               \
-  bindRec# binding =                                                                                   \
-    {- The idea here is to try and reduce the number of times registers have to be passed around -}    \
-    [|| let self ret !(o# :: Rep _o) h = $$(binding [||self||] [||ret||] [||o#||] [||h||]) in self ||] \
+#define deriveRecBuilder(_o)                                                                                               \
+instance RecBuilder _o where                                                                                               \
+{                                                                                                                          \
+  bindIterHandler# h k = [||                                                                                               \
+      let handler (linec :: Line) (colc :: Col) (c# :: Rep _o) (lineo :: Line) (colo :: Col) (o# :: Rep _o) =              \
+            $$(h (Input# [||c#||] [||linec||] [||colc||]) (Input# [||o#||] [||lineo||] [||colo||])) in $$(k [||handler||]) \
+    ||];                                                                                                                   \
+  bindIter# inp l = [||                                                                                                    \
+      let loop (line :: Line) (col :: Col) !(o# :: Rep _o) = $$(l [||loop||] (Input# [||o#||] [||line||] [||col||]))       \
+      in loop $$(line# inp) $$(col# inp) $$(off# inp)                                                                      \
+    ||];                                                                                                                   \
+  bindRec# binding =                                                                                                       \
+    {- The idea here is to try and reduce the number of times registers have to be passed around -}                        \
+    [|| let self ret h (line :: Line) (col :: Col) !(o# :: Rep _o) =                                                       \
+              $$(binding [||self||] [||ret||] [||h||] (Input# [||o#||] [||line||] [||col||])) in self ||]                  \
 };
 inputInstances(deriveRecBuilder)
 
@@ -166,10 +170,10 @@ class MarshalOps o where
   -}
   dynCont# :: StaCont# s o a x -> DynCont s o a x
 
-#define deriveMarshalOps(_o)                                          \
-instance MarshalOps _o where                                          \
-{                                                                     \
-  dynHandler# sh = [||\ (o# :: Rep _o) -> $$(sh [||o#||]) ||];        \
-  dynCont# sk = [||\ x (o# :: Rep _o) -> $$(sk [||x||] [||o#||]) ||]; \
+#define deriveMarshalOps(_o)                                                                                                    \
+instance MarshalOps _o where                                                                                                    \
+{                                                                                                                               \
+  dynHandler# sh = [||\ (line :: Line) (col :: Col) (o# :: Rep _o) -> $$(sh (Input# [||o#||] [||line||] [||col||])) ||];        \
+  dynCont# sk = [||\ x (line :: Line) (col :: Col) (o# :: Rep _o) -> $$(sk [||x||] (Input# [||o#||] [||line||] [||col||])) ||]; \
 };
 inputInstances(deriveMarshalOps)