+++ parsley-core/src/ghc-8.10+/Parsley/Internal/Backend/Machine/Types/Base.hs
@@ -19,17 +19,23 @@ module Parsley.Internal.Backend.Machine.Types.Base (
 import Control.Monad.ST                          (ST)
 import Data.STRef                                (STRef)
 import Data.Kind                                 (Type)
+import GHC.Prim                                  (Int#)
 import Parsley.Internal.Backend.Machine.InputRep (Rep)
+
+type Line = Int#
+type Col = Int#
 
 {-|
 @Handler#@ represents the functions that handle failure within a
-parser. For most of their life, handlers are represented as 
+parser. For most of their life, handlers are represented as
 `Parsley.Internal.Backend.Machine.Types.Statics.StaHandler`,
 but @Handler#@ is used at the boundaries, such as for recursion.
 
 @since 1.4.0.0
 -}
-type Handler# s o a =  Rep o          -- ^ The current input on failure 
+type Handler# s o a =  Int#           -- ^ The current line
+                    -> Int#           -- ^ The current column
+                    -> Rep o          -- ^ The current input on failure
                     -> ST s (Maybe a)
 
 {-|
@@ -39,6 +45,8 @@ feed back their result @x@ back to the caller as well as the updated input.
 @since 1.4.0.0
 -}
 type Cont# s o a x =  x              -- ^ The value to be returned to the caller
+                   -> Int#           -- ^ The current line
+                   -> Int#           -- ^ The current column
                    -> Rep o          -- ^ The new input after the call is executed
                    -> ST s (Maybe a)
 
@@ -49,8 +57,10 @@ input, an error handler in order to produce (or contribute to) a result of type 
 @since 1.4.0.0
 -}
 type Subroutine# s o a x =  Cont# s o a x  -- ^ What to do when this parser returns
-                         -> Rep o          -- ^ The input on entry to the call
                          -> Handler# s o a -- ^ How to handle failure within the call
+                         -> Int#           -- ^ The current line
+                         -> Int#           -- ^ The current column
+                         -> Rep o          -- ^ The input on entry to the call
                          -> ST s (Maybe a)
 
 {-|