+++ parsley-core/src/ghc-8.10+/Parsley/Internal/Backend/Machine/Types/Statics.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE AllowAmbiguousTypes,
              MagicHash,
+             RecordWildCards,
              TypeApplications,
              TypeFamilies #-}
 {-|
@@ -17,7 +18,7 @@ call boundaries.
 -}
 module Parsley.Internal.Backend.Machine.Types.Statics (
     -- * Handlers
-    StaHandler#, StaHandler, AugmentedStaHandler, StaHandlerCase,
+    StaHandler#, StaHandler, AugmentedStaHandler, StaHandlerCase, Input#(..), toInput, fromInput,
 
     -- ** @StaHandler@ Operations
     fromStaHandler#, fromDynHandler, staHandler#,
@@ -50,9 +51,23 @@ import Data.Kind                                       (Type)
 import Data.Maybe                                      (fromMaybe)
 import Parsley.Internal.Backend.Machine.InputRep       (Rep)
 import Parsley.Internal.Backend.Machine.LetBindings    (Regs(..), Metadata, newMeta, InputCharacteristic(..))
+import Parsley.Internal.Backend.Machine.Types.Base     (Line, Col)
 import Parsley.Internal.Backend.Machine.Types.Dynamics (DynCont, DynHandler, DynFunc)
-import Parsley.Internal.Backend.Machine.Types.Offset   (Offset(offset), same)
+import Parsley.Internal.Backend.Machine.Types.Offset   (Offset(offset), same, mkOffset, Input(..))
 import Parsley.Internal.Common.Utils                   (Code)
+
+-- Static Input
+data Input# o = Input# {
+    off#  :: Code (Rep o),
+    line# :: Code Line,
+    col#  :: Code Col
+  }
+
+fromInput :: Input o -> Input# o
+fromInput Input{..} = Input# (offset off) line col
+
+toInput :: Word -> Input# o -> Input o
+toInput u Input#{..} = Input (mkOffset off# u) line# col#
 
 -- Handlers
 {-|
@@ -62,10 +77,10 @@ on handlers, a simple form of inlining optimisation.
 
 @since 1.4.0.0
 -}
-type StaHandler# s o a = Code (Rep o) -> Code (ST s (Maybe a))
+type StaHandler# s o a = Input# o -> Code (ST s (Maybe a))
 
 mkStaHandler# :: forall o s a. DynHandler s o a -> StaHandler# s o a
-mkStaHandler# dh qo# = [||$$dh $$(qo#)||]
+mkStaHandler# dh inp = [||$$dh $$(line# inp) $$(col# inp) $$(off# inp)||]
 
 {-|
 Encapsulates a static handler with its possible dynamic origin for costless conversion.
@@ -121,7 +136,7 @@ derived from a dynamic one.
 
 @since 1.7.0.0
 -}
-augmentHandlerSta :: Maybe (Offset o) -> StaHandler# s o a -> AugmentedStaHandler s o a
+augmentHandlerSta :: Maybe (Input o) -> StaHandler# s o a -> AugmentedStaHandler s o a
 augmentHandlerSta o = augmentHandler o . fromStaHandler#
 
 {-|
@@ -133,7 +148,7 @@ if it is converted back the conversion is free.
 
 @since 1.7.0.0
 -}
-augmentHandlerDyn :: forall s o a. Maybe (Offset o) -> DynHandler s o a -> AugmentedStaHandler s o a
+augmentHandlerDyn :: forall s o a. Maybe (Input o) -> DynHandler s o a -> AugmentedStaHandler s o a
 augmentHandlerDyn c = augmentHandler c . fromDynHandler
 
 {-|
@@ -141,8 +156,8 @@ Augments a static handler with information about its captured offset.
 
 @since 1.7.0.0
 -}
-augmentHandler :: Maybe (Offset o) -> StaHandler s o a -> AugmentedStaHandler s o a
-augmentHandler c = AugmentedStaHandler c . mkUnknown
+augmentHandler :: Maybe (Input o) -> StaHandler s o a -> AugmentedStaHandler s o a
+augmentHandler c = AugmentedStaHandler (fmap off c) . mkUnknown
 
 {-|
 When the behaviours of a handler given input that matches or does not match
@@ -153,12 +168,12 @@ a handler to its argument.
 
 @since 1.7.0.0
 -}
-augmentHandlerFull :: Offset o                  -- ^ The offset captured by the creation of the handler.
+augmentHandlerFull :: Input o                   -- ^ The offset captured by the creation of the handler.
                    -> StaHandler s o a          -- ^ The full handler, which can be used when offsets are incomparable and must perform the check.
                    -> Code (ST s (Maybe a))     -- ^ The code that is executed when the captured offset matches the input.
                    -> StaHandler s o a          -- ^ The handler to be executed when offsets are known not to match.
                    -> AugmentedStaHandler s o a -- ^ A handler that carries this information around for later refinement.
-augmentHandlerFull c handler yes no = AugmentedStaHandler (Just c)
+augmentHandlerFull c handler yes no = AugmentedStaHandler (Just (off c))
   (mkFull handler
           yes
           no)
@@ -178,11 +193,11 @@ which can be used to refine the outcome of the execution of the handler as follo
 
 @since 1.7.0.0
 -}
-staHandlerEval :: AugmentedStaHandler s o a -> Offset o -> Code (ST s (Maybe a))
-staHandlerEval (AugmentedStaHandler (Just c) sh) o
-  | Just True <- same c o                   = maybe (staHandler# (unknown sh)) const (yesSame sh) (offset o)
-  | Just False <- same c o                  = staHandler# (fromMaybe (unknown sh) (notSame sh)) (offset o)
-staHandlerEval (AugmentedStaHandler _ sh) o = staHandler# (unknown sh) (offset o)
+staHandlerEval :: AugmentedStaHandler s o a -> Input o -> Code (ST s (Maybe a))
+staHandlerEval (AugmentedStaHandler (Just c) sh) inp
+  | Just True <- same c (off inp)             = maybe (staHandler# (unknown sh)) const (yesSame sh) (fromInput inp)
+  | Just False <- same c (off inp)            = staHandler# (fromMaybe (unknown sh) (notSame sh)) (fromInput inp)
+staHandlerEval (AugmentedStaHandler _ sh) inp = staHandler# (unknown sh) (fromInput inp)
 
 {-|
 Selects the correct case out of a `AugmentedStaHandler` depending on what the `InputCharacteristic` that
@@ -244,7 +259,7 @@ on continuations, a simple form of inlining optimisation.
 
 @since 1.4.0.0
 -}
-type StaCont# s o a x = Code x -> Code (Rep o) -> Code (ST s (Maybe a))
+type StaCont# s o a x = Code x -> Input# o -> Code (ST s (Maybe a))
 
 {-|
 Compared with `StaCont#`, this type also bundles the static continuation
@@ -263,7 +278,7 @@ if it is converted back the conversion is free.
 @since 1.4.0.0
 -}
 mkStaContDyn :: DynCont s o a x -> StaCont s o a x
-mkStaContDyn dk = StaCont (\x o# -> [|| $$dk $$x $$(o#) ||]) (Just dk)
+mkStaContDyn dk = StaCont (\x inp -> [|| $$dk $$x $$(line# inp) $$(col# inp) $$(off# inp) ||]) (Just dk)
 
 {-|
 Given a static continuation, extracts the underlying continuation which
@@ -292,7 +307,7 @@ on subroutines, a simple form of inlining optimisation: useful for iteration.
 
 @since 1.5.0.0
 -}
-type StaSubroutine# s o a x = DynCont s o a x -> Code (Rep o) -> DynHandler s o a -> Code (ST s (Maybe a))
+type StaSubroutine# s o a x = DynCont s o a x -> DynHandler s o a -> Input# o -> Code (ST s (Maybe a))
 
 {-|
 Packages a `StaSubroutine#` along with statically determined metadata that describes it derived from
@@ -352,5 +367,5 @@ qSubroutine :: forall s o a x rs. DynFunc rs s o a x -> Regs rs -> Metadata -> Q
 qSubroutine func frees meta = QSubroutine (staFunc frees func) frees
   where
     staFunc :: forall rs. Regs rs -> DynFunc rs s o a x -> StaFunc rs s o a x
-    staFunc NoRegs func = StaSubroutine (\dk o# dh -> [|| $$func $$dk $$(o#) $$dh ||]) meta
+    staFunc NoRegs func = StaSubroutine (\dk dh inp -> [|| $$func $$dk $$dh $$(line# inp) $$(col# inp) $$(off# inp) ||]) meta
     staFunc (FreeReg _ witness) func = \r -> staFunc witness [|| $$func $$r ||]